+++ Gruntfile.js
@@ -75,6 +75,20 @@ module.exports = function(grunt) {
                 spawn: false,
                 livereload: true
             }
+        },
+        csslint: {
+            strict: {
+                options: {
+                    import: 2
+                },
+                src: ['styles.css']
+            },
+            lax: {
+                options: {
+                    import: false
+                },
+                src: ['styles.css']
+            }
         }
     });
 
@@ -83,6 +97,7 @@ module.exports = function(grunt) {
     grunt.loadNpmTasks("grunt-contrib-watch");
     grunt.loadNpmTasks("grunt-text-replace");
     grunt.loadNpmTasks("grunt-exec");
+    grunt.loadNpmTasks('grunt-contrib-csslint');
 
     // Register tasks.
     grunt.registerTask("default", ["watch"]);